Mostly compiling version
