+++ moveit_core/CMakeLists.txt
@@ -12,7 +12,7 @@ find_package(Eigen3 REQUIRED)
 
 find_package(PkgConfig REQUIRED)
 pkg_check_modules(LIBFCL REQUIRED "fcl>=0.5.0")
-pkg_check_modules(hpp-fcl REQUIRED hpp-fcl)
+pkg_check_modules(hpp-fcl REQUIRED)
 # replace LIBFCL_LIBRARIES with full paths to the libraries
 set(LIBFCL_LIBRARIES_FULL "")
 foreach(LIBFCL_LIBRARY ${LIBFCL_LIBRARIES})
